Decompose files in .EditorConfig project to namespaces
