+++ Sources/Kysect.Configuin.EditorConfig/Analyzing/EditorConfigMissedConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.Configuin.RoslynModels;
 
-namespace Kysect.Configuin.EditorConfig;
+namespace Kysect.Configuin.EditorConfig.Analyzing;
 
 public record EditorConfigMissedConfiguration(
     IReadOnlyCollection<RoslynRuleId> StyleRuleSeverity,
