+++ Sources/Kysect.Configuin.EditorConfig/Analyzing/IEditorConfigAnalyzeReporter.cs
@@ -1,6 +1,6 @@
 ﻿using Kysect.Configuin.RoslynModels;
 
-namespace Kysect.Configuin.EditorConfig;
+namespace Kysect.Configuin.EditorConfig.Analyzing;
 
 public interface IEditorConfigAnalyzeReporter
 {
