+++ Sources/Kysect.Configuin.EditorConfig/Settings/GeneralEditorConfigSetting.cs
@@ -1,0 +1,11 @@
+﻿namespace Kysect.Configuin.EditorConfig.Settings;
+
+public record GeneralEditorConfigSetting(
+    string Key,
+    string Value) : IEditorConfigSetting
+{
+    public string ToDisplayString()
+    {
+        return $"{Key} = {Value}";
+    }
+}
