+++ Sources/Kysect.Configuin.EditorConfig/Settings/RoslynOptionEditorConfigSetting.cs
@@ -1,0 +1,5 @@
+﻿namespace Kysect.Configuin.EditorConfig.Settings;
+
+public record RoslynOptionEditorConfigSetting(
+    string Key,
+    string Value) : IEditorConfigSetting;
