+++ Sources/Kysect.Configuin.EditorConfig/Settings/RoslynSeverityEditorConfigSetting.cs
@@ -1,0 +1,7 @@
+﻿using Kysect.Configuin.RoslynModels;
+
+namespace Kysect.Configuin.EditorConfig.Settings;
+
+public record RoslynSeverityEditorConfigSetting(
+    RoslynRuleId RuleId,
+    RoslynRuleSeverity Severity) : IEditorConfigSetting;
